refactor(route): separated product routes and controllers
